chore - move utils to helpers
