+++ src/helpers/preferenceDataProcessing.ts
