+++ src/helpers/uploadDataProcessing.ts
