+++ src/helpers/uploadDataToState.ts
