+++ src/utils/index.ts
@@ -8,28 +8,15 @@ import {
   getQueryString,
   getQueryStringObj,
 } from './getQueryString';
-import { getJudgeCategory, getMeasureElement } from './measure';
 import { mergeObjInArr } from './mergeObjInArr';
-import {
-  productSizeUtil,
-  judgeProductIcon,
-  productBasicUtil,
-  productNoticeUtil,
-} from './product';
 import { updateInfo } from './updateInfo';
 
 export {
   updateInfo,
-  productSizeUtil,
-  productBasicUtil,
-  productNoticeUtil,
-  judgeProductIcon,
   deepClone,
   filterMaxPrice,
   mergeObjInArr,
   arrToString,
-  getJudgeCategory,
-  getMeasureElement,
   getQueryString,
   validatePriceRange,
   getQueryStringObj,
